log-analyzer/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2

# Data processing and analysis
pandas==2.1.2
numpy==1.26.1
python-dateutil==2.8.2

# Log parsing and pattern matching
loguru==0.7.2
regex==2023.10.3

# Machine Learning and AI
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Vector database for context storage
chromadb==0.4.15

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6
rich==13.6.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1 